requirements.txt
Django
psycopg2-binary
psycopg2
djangorestframework
djangorestframework_simplejwt
django-rest-swagger
